+++ dflib/src/test/java/org/dflib/DataFrame_Convert_ValueMapper_Test.java
@@ -4,6 +4,7 @@ import org.dflib.unit.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class DataFrame_Convert_ValueMapper_Test {
@@ -22,7 +23,7 @@ public class DataFrame_Convert_ValueMapper_Test {
     }
 
     @Test
-    public void convertColumn_ByPos() {
+    public void byPos() {
         DataFrame df = DataFrame
                 .foldByRow("a", "b")
                 .of(1, "x", 2, "y")
@@ -35,7 +36,7 @@ public class DataFrame_Convert_ValueMapper_Test {
     }
 
     @Test
-    public void convertColumn_WithNulls() {
+    public void withNulls() {
         DataFrame df = DataFrame
                 .foldByRow("a", "b")
                 .of(1, "x", 2, null)
@@ -48,7 +49,7 @@ public class DataFrame_Convert_ValueMapper_Test {
     }
 
     @Test
-    public void convertColumn_ValueMapperToDate() {
+    public void valueMapperToDate() {
         DataFrame df = DataFrame
                 .foldByRow("a")
                 .of(
@@ -65,7 +66,7 @@ public class DataFrame_Convert_ValueMapper_Test {
     }
 
     @Test
-    public void convertColumn_ValueMapperToDate_Formatter() {
+    public void valueMapperToDate_Formatter() {
         DataFrame df = DataFrame
                 .foldByRow("a")
                 .of(
@@ -78,6 +79,23 @@ public class DataFrame_Convert_ValueMapper_Test {
                 .expectHeight(3)
                 .expectRow(0, LocalDate.of(2018, 1, 5))
                 .expectRow(1, LocalDate.of(2019, 2, 28))
+                .expectRow(2, (Object) null);
+    }
+
+    @Test
+    public void valueMapperToDateTime() {
+        DataFrame df = DataFrame
+                .foldByRow("a")
+                .of(
+                        "2018-01-05T00:01:15",
+                        "2019-02-28T13:11:12",
+                        null)
+                .convertColumn("a", ValueMapper.stringToDateTime());
+
+        new DataFrameAsserts(df, "a")
+                .expectHeight(3)
+                .expectRow(0, LocalDateTime.of(2018, 1, 5, 0, 1, 15))
+                .expectRow(1, LocalDateTime.of(2019, 2, 28, 13, 11, 12))
                 .expectRow(2, (Object) null);
     }
 }