+++ dflib/src/test/java/org/dflib/SeriesType.java
@@ -1,15 +1,11 @@
 package org.dflib;
 
-import org.dflib.Series;
 import org.dflib.series.ArraySeries;
 import org.dflib.series.ColumnMappedSeries;
-import org.dflib.series.ListSeries;
 import org.dflib.series.RangeSeries;
 
-import static java.util.Arrays.asList;
-
 public enum SeriesType {
-    ARRAY, COLUMN_MAPPED, LIST, RANGE;
+    ARRAY, COLUMN_MAPPED, RANGE;
 
     public <T> Series<T> createSeries(T... data) {
         switch (this) {
@@ -19,8 +15,6 @@ public enum SeriesType {
                 return new ArraySeries<>(data);
             case COLUMN_MAPPED:
                 return new ColumnMappedSeries<>(new ArraySeries<>(data), v -> v);
-            case LIST:
-                return new ListSeries<>(asList(data));
             case RANGE:
                 return new RangeSeries<>(new ArraySeries<>(data), 0, data.length);
             default: